+++ PartnerCenterModule/PartnerCenterDirectory.psm1
@@ -15,6 +15,7 @@ $here = Split-Path -Parent $MyInvocation.MyCommand.Path
 <#
 .SYNOPSIS
 Tests to see if the specified onmicrosoft.com is available to be used for a new tenant.
+
 .DESCRIPTION
 The Get-PCDomainAvailability cmdlet tests to see if the specified tenant domain (onmicrosoft.com) is available to create a new tenant.
 
@@ -189,9 +190,9 @@ function Add-PCCustomerRoleMember
 
 <#
 .SYNOPSIS
-
+TODO
 .DESCRIPTION
-
+The Remove-PCCustomerRoleMember cmdlet
 .PARAMETER SaToken 
 Specifies a security token for authenticating and executing the cmdlet.